[MISC] Move print_version to format_base_help since it does not contain specific information anymore.
