+++ include/seqan3/argument_parser/detail/format_help.hpp
@@ -20,7 +20,6 @@
 #include <seqan3/argument_parser/detail/format_base.hpp>
 #include <seqan3/argument_parser/detail/terminal.hpp>
 #include <seqan3/core/detail/test_accessor.hpp>
-#include <seqan3/version.hpp>
 
 namespace seqan3::detail
 {
@@ -194,27 +193,6 @@ protected:
         print_text(desc, layout.rightColumnTab);
 
         prev_was_paragraph = false;
-    }
-
-    //!\brief Prints the version information to std::cout.
-    void print_version()
-    {
-        std::string const version_str = std::to_string(SEQAN3_VERSION_MAJOR) + "." +
-                                        std::to_string(SEQAN3_VERSION_MINOR) + "." +
-                                        std::to_string(SEQAN3_VERSION_PATCH);
-
-        // Print version, date and url.
-        print_section("Version");
-        print_line(in_bold("Last update: ") + meta.date, false);
-        print_line(in_bold(meta.app_name + " version: ") + meta.version, false);
-        print_line(in_bold("SeqAn version: ") + version_str, false);
-
-        if (!empty(meta.url))
-        {
-            print_section("Url");
-            print_line(meta.url, false);
-            std::cout << "\n";
-        }
     }
 
     //!\brief Prints a help page footer to std::cout.
